# swagger/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "varsion": "0.0.1",
    "title": "DAS SDL API server",
    "description": "DAS SDL 서버 API swagger 문서",
    "version": "1.0.0"
  },
  "host": "localhost:3200",
  "basePath": "/das",
...
  "paths": {
...
    "/prc/prcInfo": {
      "get": {
        "tags": [
          "prc"
        ],
        "summary": "프로젝트 특정 버전 프로세스 단계별 조회",
        "description": "사용자가 프로젝트 특정 버전 프로세스 단계별 리스트를 조회한다",
        "parameters": [
          {
            "name": "prj_id",
            "in": "query",
            "type": "number",
            "required": true,
            "example": 1,
            "description" : "조회 대상 프로젝트 아이디"
          },
          {
            "name": "version_number",
            "in": "query",
            "type": "string",
            "required": true,
            "example": "v0.0.1",
            "description" : "조회 대상 프로젝트 버전"
          },
          {
            "name": "step_number",
            "in": "query",
            "type": "number",
            "required": true,
            "example": 1,
            "description" : "조회 대상 프로젝트 버전 스텝 번호"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "resultCode": {
                  "type": "number",
                  "example": 200
                },
                "resultMsg": {
                  "type": "string",
                  "example": "프로세스 단계 조회 성공"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/prc/addLstc": {
      "post": {
        "tags": [
          "prc"
        ],
        "summary": "프로젝트 특정 버전 프로세스 리스트 체크파일을 추가",
        "description": "사용자가 프로젝트 특정 버전 프로세스 리스트 체크파일을 추가한다",
        "consumes": ["multipart/form-data"],
        "parameters": [
          {
            "name": "prj_id",
            "in": "formData",
            "type": "number",
            "required": true,
            "example": 1,
            "description" : "프로젝트 아이디"
          },
          {
            "name": "version_number",
            "in": "formData",
            "type": "string",
            "required": true,
            "example": "v0.0.1",
            "description" : "프로젝트 버전"
          },
          {
            "name": "file",
            "in": "formData",
            "type": "file",
            "required": true,
            "description" : "리스트 체크 파일(단일)"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "resultCode": {
                  "type": "number",
                  "example": 200
                },
                "resultMsg": {
                  "type": "string",
                  "example": "리스트 체크 파일 등록 성공"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
    "/prc/addCmt": {
      "post": {
        "tags": [
          "prc"
        ],
        "summary": "프로젝트 특정 버전 프로세스에 코멘트 등록",
        "description": "사용자가 프로젝트 특정 버전 프로세스에 코멘트를 등록한다",
        "consumes": ["multipart/form-data"],
        "parameters": [
          {
            "name": "prj_id",
            "in": "formData",
            "type": "number",
            "required": true,
            "example": 1,
            "description" : "프로젝트 아이디"
          },
          {
            "name": "version_number",
            "in": "formData",
            "type": "string",
            "required": true,
            "example": "v0.0.1",
            "description" : "프로젝트 버전"
          },
          {
            "name": "step_number",
            "in": "formData",
            "type": "number",
            "required": true,
            "example": 1,
            "description" : "프로세스 번호"
          },
          {
            "name": "rgst_user_id",
            "in": "formData",
            "type": "number",
            "required": true,
            "example": 1,
            "description" : "코멘트 등록자 아이디"
          },
          {
            "name": "comment_description",
            "in": "formData",
            "type": "string",
            "required": true,
            "example": "어쩌구 저쩌구 ㅇㅇㅇ",
            "description" : "코멘트 내용"
          },
          {
            "name": "files",
            "in": "formData",
            "type": "array",
            "items" : {
              "type" : "file"
            },
            "description" : "코멘트 등록 파일(다중)"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "resultCode": {
                  "type": "number",
                  "example": 200
                },
                "resultMsg": {
                  "type": "string",
                  "example": "코멘트 등록 성공"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/prc/updtCmt": {
      "put": {
        "tags": [
          "prc"
        ],
        "summary": "코멘트 수정",
        "description": "사용자가 코멘트를 수정한다.",
        "consumes": ["multipart/form-data"],
        "parameters": [
          {
            "name": "comm_id",
            "in": "formData",
            "type": "number",
            "required": true,
            "example": "1",
            "description" : "변경대상 코멘트 아이디"
          },
          {
            "name": "comment_description",
            "in": "formData",
            "type": "string",
            "example": "어쩌구저쩌구",
            "description" : "변경하려는 내용"
          },
          {
            "name": "del_file_id",
            "in": "formData",
            "type": "array",
            "items" : {
              "type" : "number"
            },
            "example": 1,
            "description" : "삭제할 코멘트 파일 번호(다중)"
          },
          {
            "name": "files",
            "in": "formData",
            "type": "array",
            "items" : {
              "type" : "file"
            },
            "description" : "변경할 프로젝트 파일 (다중)"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "resultCode": {
                  "type": "number",
                  "example": 200
                },
                "resultMsg": {
                  "type": "string",
                  "example": "코멘트 수정 완료"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/prc/delCmt": {
      "delete": {
        "tags": [
          "prc"
        ],
        "summary": "코멘트 삭제",
        "description": "사용자가 코멘트를 삭제한다.",
        "consumes": ["multipart/form-data"],
        "parameters": [
          {
            "name": "comm_id",
            "in": "formData",
            "type": "number",
            "required": true,
            "example": "1",
            "description" : "변경대상 코멘트 아이디"
          },
          {
            "name": "comment_description",
            "in": "formData",
            "type": "string",
            "example": "어쩌구저쩌구",
            "description" : "변경하려는 내용"
          },
          {
            "name": "del_file_id",
            "in": "formData",
            "type": "array",
            "items" : {
              "type" : "number"
            },
            "example": 1,
            "description" : "삭제할 코멘트 파일 번호(다중)"
          },
          {
            "name": "files",
            "in": "formData",
            "type": "array",
            "items" : {
              "type" : "file"
            },
            "description" : "변경할 프로젝트 파일 (다중)"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "resultCode": {
                  "type": "number",
                  "example": 200
                },
                "resultMsg": {
                  "type": "string",
                  "example": "코멘트 삭제 완료"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/prc/lstnRgst": {
      "post": {
        "tags": [
          "prc"
        ],
        "summary": "프로세스 최종 정보 등록",
        "description": "사용자가 프로세스의 최종 정보를 등록한다",
        "consumes": ["multipart/form-data"],
        "parameters": [
          {
            "name": "prj_id",
            "in": "formData",
            "type": "number",
            "required": true,
            "example": 1,
            "description" : "등록 대상 프로젝트 아이디"
          },
          {
            "name": "version_number",
            "in": "formData",
            "type": "string",
            "required": true,
            "example": "v0.0.1",
            "description" : "등록 대상 프로젝트 번호"
          },
          {
            "name": "step_number",
            "in": "formData",
            "type": "number",
            "required": true,
            "example": 1,
            "description" : "등록 대상 프로젝트 스텝 번호"
          },
          {
            "name": "step_lnk",
            "in": "formData",
            "type": "string",
            "example": "https://naver.com",
            "description" : "등록하려는 링크"
          },
          {
            "name": "step_description",
            "in": "formData",
            "type": "string",
            "example": "어쩌구저쩌구 왈왈",
            "description" : "등록하려는 내용"
          },
          {
            "name": "files",
            "in": "formData",
            "type": "array",
            "items" : {
              "type" : "file"
            },
            "description" : "등록하려는 파일 (다중)"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "resultCode": {
                  "type": "number",
                  "example": 200
                },
                "resultMsg": {
                  "type": "string",
                  "example": "프로젝트 최종 정보 등록 성공"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
  }
}